Revert "feat: Refactor register and scanning files"
This reverts commit c4314eb1c0a1eefec4bee542b374f85eb32870ff.
