+++ package.json
@@ -50,7 +50,6 @@
     "matter-js": "^0.20.0",
     "multer": "^1.4.3",
     "next": "^15.1.7",
-    "next-auth": "^4.24.11",
     "next-connect": "^0.11.0",
     "next-fonts": "^1.5.1",
     "next-pwa": "^5.6.0",
@@ -73,7 +72,6 @@
     "swiper": "^11.1.1",
     "twilio": "^5.3.3",
     "webpack": "^5.89.0",
-    "yup": "^1.6.1",
     "zustand": "^4.5.5"
   },
   "devDependencies": {
